openfst-1.2.4.tar.gz r1
SHA-256: 1e5d84591f33ec34af52201efc7d73731d7453d64ca0701688bb55d49e83a080

+++ src/extensions/pdt/pdtreplace.cc
@@ -50,13 +50,13 @@ int main(int argc, char **argv) {
   typedef int64 Label;
   typedef pair<Label, const s::FstClass* > FstTuple;
   vector<FstTuple> fst_tuples;
-  Label root = atoi(argv[2]);
+  Label root = atoll(argv[2]);
   fst_tuples.push_back(make_pair(root, ifst));
 
   for (size_t i = 3; i < argc - 1; i += 2) {
     ifst = s::FstClass::Read(argv[i]);
     if (!ifst) return 1;
-    Label lab = atoi(argv[i + 1]);
+    Label lab = atoll(argv[i + 1]);
     fst_tuples.push_back(make_pair(lab, ifst));
   }
